+++ cppcache/src/CacheableObjectPartList.hpp
@@ -147,7 +147,7 @@ class CacheableObjectPartList : public Cacheable {
    */
   virtual int8_t DSFID() const;
 
-  virtual uint32_t objectSize() const;
+  virtual size_t objectSize() const;
 };
 
 }  // namespace client
